feat: add failurePolicy and scalingTimeout to CRD spec
scalingTimeout defines how long a cluster is allowed to stay in scaling mode.
If it exceeds the limit the cluster will go into a failed state.

failurePolicy defines how to handle a failed cluster. By default it will
retain the existing pods. This can be set to Recreate to trigger re-creation
of an etcd cluster.

These two fields work together to ensure that if a cluster creation times out
it will be re-created from scratch.

Signed-off-by: Moritz Johner <Moritz.Johner@form3.tech>